data-extraction-pipeline/requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
camelot-py==0.7.3
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
et-xmlfile==1.0.1
idna==2.9
jdcal==1.4.1
numpy==1.18.2
opencv-python==4.2.0.32
openpyxl==3.0.3
pandas==1.0.3
pdfminer.six==20200124
pkg-resources==0.0.0
pycryptodome==3.9.7
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sortedcontainers==2.1.0
soupsieve==2.0
urllib3==1.25.8
